Creado un selector de imágenes para subir fotos por JS

+++ src/modules/daybook/views/EntryView.vue
@@ -7,12 +7,14 @@
     </div>
 
     <div>
+      <input type="file" @change="onSelectedImage" ref="imageSelector" v-show="false" accept="image/png, image/jpeg">
+
       <button v-if="entry.id" class="btn btn-danger mx-2" @click="onDeleteEntry">
         Borrar
         <i class="fa fa-trash-alt"></i>
       </button>
 
-      <button class="btn btn-dark">
+      <button class="btn btn-dark" @click="onSelectImage">
         Subir foto
         <i class="fa fa-upload"></i>
       </button>
@@ -27,18 +29,21 @@
 
   <Fab icon="fa-save" @on-click="saveEntry"/>
 
-  <img class="img-thumbnail" src="../assets/vue.jpg" alt="entry-picture">
+  <img v-if="localImage" class="img-thumbnail" :src="localImage" alt="entry-picture">
 </template>
 
 <script>
 import { defineAsyncComponent } from 'vue'
 import { mapGetters, mapActions } from 'vuex'
 import getDate from '@/modules/daybook/helpers/getDate.js'
+import Swal from 'sweetalert2'
 
 export default {
   data() {
     return {
-      entry: null
+      entry: null,
+      localImage: null,
+      file: null,
     }
   },
 
@@ -88,11 +93,16 @@ export default {
         entry = this.getEntryById(this.id)
         if(!entry) return this.$router.push({name: 'no-entry'})
       }
-
       this.entry = entry
     },
 
     async saveEntry() {
+      new Swal({
+        title: 'Espere por favor',
+        allowOutsideClick: false,
+      })
+      Swal.showLoading()
+
       if(this.entry.id) {
         await this.updateEntry(this.entry)
       }
@@ -101,11 +111,46 @@ export default {
         this.$router.push({name: 'entry', params: {id}})
       }
 
+      Swal.fire('Guardado', 'Entrada registrada con éxito', 'success')
+
     }, 
 
     async onDeleteEntry() {
-      await this. deleteEntry(this.entry)
-      this.$router.push({name: 'no-entry'})
+      const result = await Swal.fire({
+        title: '¿Está seguro?',
+        text: 'Una vez borrado, no se puede recuperar',
+        showDenyButton: true,
+        confirmButtonText: 'Si, estoy seguro'
+      })
+      
+      if(result.isConfirmed) {
+        new Swal ({
+          title: 'Espere por favor',
+          allowOutsideClick: false,
+        })
+        Swal.showLoading()
+        await this. deleteEntry(this.entry)
+        this.$router.push({name: 'no-entry'})
+
+        Swal.fire('Eliminado', 'Entrada eliminada con éxito', 'success')
+      } 
+    },
+
+    onSelectedImage(event) {
+      const file = event.target.files[0]
+      if(!file) {
+        return
+      }
+
+      this.file = file
+
+      const fr = new FileReader()
+      fr.onload = () => this.localImage = fr.result
+      fr.readAsDataURL(file)
+    }, 
+
+    onSelectImage() {
+      this.$refs.imageSelector.click()
     }
   },
